requirements.txt
streamlit
PyPDF2
pypdf
numpy
google-generativeai>=0.7.0
python-dotenv
pydantic